refactor: support multiple languages per runner (#161)
This is a bit of a hack at the DB layer, in that we're storing multiple
values in a single field (eg. ":go:kotlin:"), but the number of active
runners should be relatively small (low 1000's max) and it vastly
simplifies the refactor. We can reassess later if it becomes an issue.

Also extended the Runner to support initialising the user code root from
a template directory. This allows the Dockerfile to seed the template
root with eg. `ftl-runtime.jar`.

To start a Docker-based Runner:

```
make docker-runner
docker run -p 8899:8899 --rm -it --platform linux/amd64 ftl-runner
```