Return logs without a request key (#229)
With the new event setup, we were previously only returning log events
that had a `request_key`. But many of the log events happen outside of a
request context. For example, a new deploy won't have an associated
`request_key`.

This PR makes request_key optional so we can get logs with and without
the key.