fix: case where a runner ends up reserved without a timeout (#393)
I believe this was being triggered when a controller creates a
reservation transaction, issues the reservation to a runner, then dies
before being able to complete the transaction. This leaves the runner in
a state where it is telling controllers it's reserved, without having
set a reservation_timeout, as the only place that was previously set was
during the reservation transaction. The UpsertRunner calls were not
setting this.

To work around this, and another potential issue, I've added two
triggers:

1. When a runner is set to the reserved state with a NULL
reservation_timeout, set it to a default of 2m.
2. Update runner.module_name whenever runner.deployment_id is set or
unset.

Fixes #392