fix: optimise console timeline query (#631)
Before (~3s):
https://www.pgexplain.dev/plan/02fdabad-f469-45ae-a493-6757083ad307
After (7ms):
https://www.pgexplain.dev/plan/f92f1325-11a7-4166-862d-7490e80adb71

The issue lay in `JOIN e.deployment_id = d.id`. For reasons that are
still unknown, replacing this with a static `WHERE e.deployment_id IN
(x, y, z)` completely fixes the query.

This makes the Go code quite a bit uglier, but until (or if) we figure
out how to make the single query fast this will have to suffice, as I
spent hours trying to make it work to no avail.