fix: ensure Go runtime uses the correct field names without "json" tags (#706)
FTL's schema requires all fields be in "lowerCamelCase", but the
go-runtime has been encoding fields with the literal field name, usually
resulting in "UpperCamelCase". This change adds a custom encoder
(because there is seemingly no existing library that will do this).

Decoding should be fine because Go's JSON decoder is case-insensitive.