feat: embed a JS interpreter for `ftl schema generate` templates
Embed the goja JS interpreter and extend the `flt schema generate`
templates to load a template.js file from the template directory. Each
top-level function in the JavaScript file will be exposed in the Go
template function map.

Additionally the normal `console.log` et al functions are available, and
a `repr(v)` function for dumping a Go representation of a value in the
template.