feat: allow arbitrary types in requests/responses (#763)
This is a prerequisite for PubSub, but we've always wanted to loosen
this restriction anyway. It was only in place temporarily until we get
concurrent deployed versions working, but I think we can just guide
people until that point.

There are still restrictions for `@Ingress` verbs in that the request
must still be a data structure, because the routing needs something to
map the path parameters and query parameters onto.