feat: add support for different content types (#835)
Basic support for different content-types on HTTP responses. This will
need to be extended in the future, but is setup now to support `text/`
types. This will definitely need more thought in the future.

This allows us to return HTML content from an HTTP ingress

```go
return builtin.HttpResponse[string]{
	Status:  200,
	Headers: map[string][]string{"Content-Type": {"text/html; charset=utf-8"}},
	Body:    "<html><body><h1>HTML Page From FTL 🚀!</h1></body></html>",
}, nil
```

![Screenshot 2024-01-25 at 10 12
07 AM](https://github.com/TBD54566975/ftl/assets/51647/acb413f9-5be3-4c5c-868b-dda155512f44)