fix: handle data refs to type parameters
I originally intended to replace `DataRef` references to type
parameters, with concrete `TypeParameter` instances, but this ended up
conflating the parameter declaration and the reference. Instead, it's
now left as a `DataRef`, and I've made `TypeParameter` no longer
implement `Type`.

Note that it being a `DataRef` doesn't make much sense either, because
it's not a data structure, but that's part of a large issue where data
references can also refer to verbs within the schema parsing. The
correct solution here is to use an untyped `Ref` everwhere and resolve
these to the correct reference types when validating.