fix: databasetesting.CreateForDevel() now tries to connect for 10s (#886)
There seems to be a race between when the Docker PG healthcheck passes,
and when the Docker port becomes available:

trace: exec: cd . && docker inspect --format \{\{.State.Health.Status}}
ftl-db-1
    debug: Postgres DSN: ***localhost:54320/ftl?sslmode=disable
ftl: error: failed to connect to `host=localhost user=postgres
database=`: failed to receive message (unexpected EOF)

It's not clear if that's a deficiency in Docker, or perhaps pg_isready
is only checking the Unix socket and not the TCP port, and there's some
delay between the two becoming available. Either way, this should solve
it.