fix: template directories must be ignored by Go
`|` is not supported in filenames retrieved by the Go toolchain, as
Windows doesn't allow it in filenames. As we're zipping these files,
this isn't a concern for us, but it's still prevented by Go. To work
around this we create a dummy `go.mod` file which Go interprets as a
separate module and thus ignores.

Fixes #991