chore: fix for local dev rebuilding ftl-runtime.jar on every run
This was occurring because frontend/local.go was calling `npm install`
explicitly on every run, resulting in changes to the frontend that
invalidated some of the dependencies of the JAR file.

The fix is to use `bit` to build the frontend as well, resulting in its
dependency cache being correct.