feat: only do autofmt on main (#1156)
for issue: https://github.com/TBD54566975/ftl/issues/1075

Not sure of best way to test - I may try it on another branch perhaps? 

edit: tested it by having it push to this branch as if it was main, and
it did what I expected. Doesn't need the FTL_AUTOFIX_TOKEN anymore, and
won't trigger a rebuild (flip side of that is if it does have a rebuild,
there won't be a build run against it - is that suitable @alecthomas ?)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>