chore: use mdcode to avoid bitrot in README (#1172)
[mdcode](https://github.com/szkiba/mdcode) is a tool that allows code
snippets in Markdown files to be tested. This PR uses it to test that
the instructions in the README don't bitrot.

We can/should use this to ensure the instructions in our future
documentation stay current.

This required changes to how the controller reports active deployments,
some minor refactoring, and adding functionality to the controller that
allows it to report readiness only once a set of modules are deployed.