refactor: rename Singleton to Map
And change its purpose slightly such that it is used to map FTL
resource handles to another type.

Rationale: the main goal of introducing the Singleton type was to allow
custom transformation of FTL resources, and this more closely reflects
that goal. I didn't thin of doing it in this way until later.