Resolve Header Duplication
It has been found that when version 1.16.0 of connectrpc is used, that there can be duplicate headers present which results in malformed headers when using FTL in an environment where it communicates through an Envoy sidecar.

This change downgrades to 1.14.0 as that is a known good version, and it also prevents the renovate bot from updating this dependency.