feat: expose leases to user code (#1406)
[Leases](https://hackmd.io/@ftl/Sym_GKEb0) provide exclusive access to a
resource identified by a unique key with a timeout. If the timeout
expires due to the lease holder terminating, the lease will be released.

Usage is straightforward:

```go
lease, err := ftl.Lease(ctx, 10*time.Second, "update", "user", "123")
if err != nil {
        return err
}
defer lease.Release()

// ...
```

Note that leases are currently not tracked in the schema.