refactor: add user-accessible transactions to the DAL
Up until now all transactions were hidden behind DAL methods, but in
order to cleanly separate the core of async calls from the type-specific
post-execution callbacks, we'll need to pass an open DAL transaction to
the callback.

eg. when an FSM transition async call completes there will be a callback
that will update the executing FSM instance with the destination state.