Fix Header Duplication
When using connectrpc to perform calls, duplicate headers are introduced if the response object is returned directly by the runner or the controller. This causes bugs when FTL is deployed to an environment that does not work when there are duplicate headers, such as an environment where envoy is used to facilitate traffic.

The fix for this is to instantiate a new response object to return from the Call functions.

For more information on the envoy bug, see the issue linked below.
https://github.com/envoyproxy/envoy/issues/9019

The fix for this took place after reaching out to the Buf team on slack
https://bufbuild.slack.com/archives/CRZ680FUH/p1716344335783839