fix: refactor manager.go to use provider key instead of MutableProviders, adds `--opvault=VAULT` controller flag (#1604)
This change is a precursor to #1565. It removes the notion of
`MutableProviders` and requires a provider be specified when calling
mutable functions in the manager. It also allows a 1Password vault to be
set at the controller & main.go levels by using `--opvault=VAULT`. To
operate on 1P secrets via CLI, continue to use `ftl secret ...
--op=VAULT`, but this should eventually change to a bool flag.