feat: generate ftl-project.toml in the Git root if it DNE (#1648)
Fixes https://github.com/TBD54566975/ftl/issues/1639

The current implementation ignores any `--config` flag values, but let
me know if we should default to creating the last _specified_ config
whenever one's provided. For that, I'd just move this code to the
resolver and add an `r.Config` check. I decided to minimize it purely to
operating on the default path because the specified paths may not
actually be at the Git root, which could make the new behavior a bit
more complex to comprehend. That said, there could be a good reason to
use the flag value that I didn't think of.