fix: introduces runner support for a dynamically refreshing module context (#1822)
Introduces a `ModuleContext` wrapper that supplies the most recent one
returned by `GetModuleContext` which now streams results. Currently the
result stream is simply a stream of one item - my next change will
update the controller to send changes through the stream.

The startup code waits for the the first value to be returned from the
stream (currently hard code 5 seconds timeout but I suspect a more
reasonable strategy will be suggested in review) so that a
`ModuleContext` is always available to the Runner.