feat: introduce `types.ftl.go` to register sumtypes, which fixes encoding in unit tests (#1909)
Fixes https://github.com/TBD54566975/ftl/issues/1577

This change introduces a new generated file `types.ftl.go` in the
user-defined package itself to make sure sumtype registrations get
called when running unit tests.

Notes:
* This change also includes a subpackage `subpackage` in the test data
that is not yet unit tested. Unit tests should be added (identical to
the parent package's unit tests) after
https://github.com/TBD54566975/ftl/issues/1903 is resolved. Some work
will also need to be done to extend the `types.ftl.go` generation to
working on subpackages.
* This works by extending the existing logic operating on
`mainModuleContext`, which means that code now generates not only just
the literal `main` package, but also the conceptual _main_ package's
additional artifacts. I couldn't come up with a name that would be
better, so it's left as-is in this PR, but please let me know if you
think of something better.