fix: proactively coordinate leader/followers (#2028)
fixes #2011

Cause: 
- Coordinator only does it coordination logic when something calls
`Get()`
- If leader/follower coordinator's `Get()` function was not called then
an expired leader/follower would not change
- ASM follower syncs with what it expects to be the leader even if that
leader is not alive. Even if we get a new leader, the follower would not
be swapped out for a valid one until `Get()` was called. So it would
churn away failing to sync.

Fix:
- Proactively call `Get()` periodically so that coordination occurs even
if no calls to Get() would otherwise happen.