fix: avoid insertion key dupls in `module_secrets` table by upserting on conflict (#2105)
Addresses part of https://github.com/TBD54566975/ftl/issues/2084

Postgresql way to handle upserts:
https://www.postgresqltutorial.com/postgresql-tutorial/postgresql-upsert/

This PR fixes the issue for secrets since that's causing issues for the
client team, but not for config, which has the same issue. I confirmed
the test fails before this change but passes with it.