feat: bucketize latency attribute (#2269)
This PR bucketizes the latency _attribute_ (NOT the latency _metric_!
That would be very bad!)

For better filtering capabilities in datadog, we need to bucket our
high-cardinality numeric attributes. This uses log base8 because:

1. That seems to give the most useful results for current production
numbers we're seeing in DD - good functional granularity for analysis
purposes without being so granular that it's annoying to sort through
the buckets. Current values range from 62 to 20279, concentrated in the
100-1000 range.
2. Base 8 bucket labels are all powers of 2, so it's relatively easy for
our brains to skim

Sample output:
```
-> ftl.async_call.time_since_scheduled_at_ms.bucket: Str([8,64))
```

Issue: https://github.com/TBD54566975/ftl/issues/2194