feat: propagate otel spans across pubsub (#2287)
This adds an `otel_context` column to both `topic_events` and
`async_calls`.
- on publish, the current otel context is stored in `topic_events`
- `otel_context` is propagated to the `async_calls` row on dequeue and
retries
- on execution of a row in `async_calls`, the current otel trace is
replaced with `otel_context`

Pros:
- we can see related calls all in the same trace
- pattern can be reused for all async calls (i.e. fsm)

Cons:
- storing otel specific context in the db, feels a bit gross
- there is no root span tying the retries / subscriptions together,
since the root would be the controller lifecycle. screenshot below.

An alternative would be to persist the `requestKey` in the tables
instead of `otel_context` and attach that to subscribers / retries, and
use that as a filter in DD, etc.

<img width="774" alt="Screenshot 2024-08-07 at 12 59 43 PM"
src="https://github.com/user-attachments/assets/2cd61392-0ff8-4032-8ca4-20e4aec033de">

Closes #2258