feat: add support for code and SQL based database migrations
We'll need to be doing data migrations quite a bit in the near future,
and while at some point we will need to do proper live migrations with
all that this entails, for now it will be more expedient to accept some
downtime. With that in mind, this PR adds a function for applying a mix
of SQL and code-based migrations to a database.

An example of where this will be used is our encrypted table columns,
which are currently JSONB with a base-64 encoded JSON envelope. We want
to move to raw encrypted bytes. We might also need to reencrypt some
data with new keys soon. etc.