addresses the pubsub event synchronization issues resulting in the race condition.
issue #1
the waitgroup used to await event processing completion is not synchronized with event handling completion. The waitgroup has its count synchronously incremented before the event gets published but the down tick is not explicitly synchronized with handler completion. The down tick occurs in `handlePubSubEvent` in response to a `publishEvent` signal - this signals the intention to distribute an event on a topic not the completion. The `subscriptionDidConsumeEvent` is the right event to target. Given the event is dispatched via a goroutine the race condition may or may not trigger failure depending on competition execution timings.

issue #2
event synchronization needs to consider the subscription topology in order to account for non-linear event propagation networks

remaining issue
synchronous event propagation from the subscriber verb is covered but not asynchronous propagation (e.g. a scenario where a verb spawns a goroutine that delays before writing an event to another topic might fire the event after the waitgroup reaches zero)