feat: Initial Java runtime implementation (#2318)
This is still very much a work in progress, however it contains a lot of
basic functionality.

So far this includes support for:

- Verb invocations
- HTTP ingress
- Cron
- Topics and Subscriptions
- Basic testing of Verbs

The existing Kotlin example has been migrated over to the new approach.
At the moment the module is called Java even though it supports both, in
future this will provide a base layer of functionality with some small
language dependent features in separate Java/Kotlin modules.