refactor: decouple conflicting ModuleContext requirements (#2376)
The ModuleContext was designed to be an abstract data model in the
Controller for the resources required by a module, but along the way it
started to be used for storing DB connections for use by the go-runtime.
This change cleanly separates those requirements so that the go-runtime
is entirely responsible for creating new connections from the DSN
provided by the ModuleContext.

I think there's a bit more work to be done here, in that the
ModuleContext knows about testing in the go-runtime, which it really
shouldn't, but this will unblock #2373 for now.