feat: add directive to enable lenient json (#2377)
fixes #2132

adds support for `ftl:encoding` directive. the default `ftl:encoding`
type is json and presently supports only the `lenient` option.

`//ftl:encoding lenient` or `//ftl:encoding json lenient` (which are
equivalent) can annotate an ingress verb. when an ingress verb is
annotated with this, it will accept requests with unknown/extra fields
and ignore them.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>