feat: allow FSMs to specify the next event to send (#1723)
closes #1664
closes #2334

This is particularly useful when an FSM state itself knows which state
it should be in next, for example transitioning a payment to "voided" if
an error occurs communicating with an external vendor.

- Adds `fsm.Next(ctx, event)` to the go runtime
    - Only one next transition is allowed per fsm instance
- When a transition completes, the next transition is queued up if it
exists
- When a transition returns an error, we wipe the next transition so
that the retry attempt can set the next transition again

In another PR we will move away from using `ftl.Next()` to avoid the
reference cycle issue.

---------

Co-authored-by: Matt Toohey <mtoohey@block.xyz>