feat: change how runner provisioning works (#2590)
This changes the way runners are provisioned, and how runners are
allocated. Runners are
now spawned knowing exactly which kube deployment they are for, and will
always
immedatly download and run that deployment.

For kubernetes environments replicas are controlled by creating a kube
deployment
for each FTL deployment, and adjusting the number of replicas.

For local scaling we create the runners directly for deployments as
required.

This also introduces an initial kubernetes test.