feat: Provisioner Plugin POC (#2716)
This adds a Cloudformation based Plugin for provisioning databases. Some
things to note:
- This is a POC and not intended for production
- Still missing Plan API
- We are passing requited underlying subnet group id as a resource
property of a dependent "FTL Cluster" resource. This needs to be
eventually populated by FTL bootstrap process, when creating the
cluster.
- `dev` runner is just a helper to try this out locally. It needs to
evolve something much more generic and useful in the future.

This creates a single Cloudformation Stack / module, with all module
resources in it

![Screenshot 2024-09-18 at 11 38
02 AM](https://github.com/user-attachments/assets/26d70565-fbaf-440b-941c-4d098983ca69)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>