fix: scheduled tasks should not derive names (#2840)
Singleton scheduled tasks were broken due to the job functions being
wrapped. This was causing the derived name for each job to be the same,
which meant they all competed for the same lease.

Since we recently added a name for each of the scheduled tasks, this PR
passes those names along to the scheduler rather than relying on a
derived name.