feat: Provisioner for FTL modules (#3050)
Apologies, this grow a bit bigger than I hoped for.

Major parts in this PR:
- a new in memory provisioner for creating modules by calling
`ftl-controller`
- an in memory resource graph to represent that modules can depend on
databases. The direct dependants are passed to the provisioner when
provisioning, with outputs filled in from previous runs. Eventually,
this graph will be saved to DB

Next: pass the new database DSNs to the controller at deployment.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>