chore: Remove redundant otel collector (#3087)
# Summary
This PR scraps a redundant otel collector in favor of the one that's
included in the [otel-lgtm](https://github.com/grafana/docker-otel-lgtm)
container we use for grafana. Also took an _initial_ crack at
documenting the metrics that FTL exports.


> [!NOTE]
> will include docs for traces and logs in subsequent PRs


# Rationale
I noticed that the docker image we're using for grafana actually runs a
full blown observability stack that grafana provides for development,
demo, and testing environments. this stack includes an otel collector as
well so figured it made sense to ditch the solo collector


# Changes
* scrapped `otel-collector` from `docker-compose`
* renamed `grafana` service to `otel-lgtm` to more accurately reflect
what's running
* disabled the annoying grafana login form
* renamed `just grafana` to `just observe`
* renamed `just grafana-stop` to `just observe-stop`
* updated `CONTRIBUTING.md`
* added `observability/metrics` docs

Exported metrics can be seen like so:

> [!NOTE]
> apologies for the stuttering mouse movement. my arm was cat trapped


https://github.com/user-attachments/assets/e51c0fa3-01af-4c6f-93d3-37fe09698cda